Update patch set 1
Patch Set 1:

A lot of people use forks of the CM11.
In some of these forks it is not possible to use CMAccounts but they are using the whispersystem app.
Therefore it is not possible to delete the number from the CM servers by deleting the CMAccount - there is none.

Could you please add the unregister to the whispersystem app?

Patch-set: 1
Reviewer: Egon Gupta <9118@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0