Update patch set 1
Patch Set 1:

Correct.  I'm going to be working on a solution to unregister if you have wiped or are no longer running CM sometime next week.  It will be a web-based unregistration process that requires you reverify your phone number by sending it a text message with a verification code.

Patch-set: 1