Update patch set 1
Patch Set 1: Verified+1 Code-Review-1

Let's hold off on pulling this into stable.  We need to get a build cut in the next hour, and I don't think that is enough time to get this fully tested.  We have not been able to repro the issue that this change is in place to deal with after the last round of bug fixes.

Patch-set: 1
Reviewer: Chris Soyars <1@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=-1