Update patch set 1
Patch Set 1:

This would work for signing it with the cyngn-app key. However, if we are going to grant this key special permissions, we should not use the generic cyngn-app key, we should create a new key for this use case. If this is what we are doing, please create a ticket to request a new key to be issued - you can see https://jira.cyngn.com/browse/RM-115 as an example.

Patch-set: 1
Label: Code-Review=0